Add commands to api and clarify existing cmds
Added the command number to the name of each so they appear
in that order on the website. I also removed all commands
from README which I have added to api.js. I am aware of
all the end-of-line whitespace, descriptions are formatted
with markdown, so a double space forces a newline, which is
the reason for the excessive whitespace.